refactor(Coupon): CouponQueryRepository 패키지 위치 변경
