+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/repository/CouponQueryRepository.java
@@ -1,4 +1,4 @@
-package shop.bluebooktle.backend.coupon.repository.query;
+package shop.bluebooktle.backend.coupon.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
