+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/repository/CouponTypeRepository.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import shop.bluebooktle.backend.coupon.entity.CouponType;
-import shop.bluebooktle.backend.coupon.repository.query.CouponQueryRepository;
 
 public interface CouponTypeRepository extends JpaRepository<CouponType, Long>, CouponQueryRepository {
 	Optional<CouponType> findByName(String couponName);
