+++ backend-server/src/main/java/shop/bluebooktle/backend/coupon/repository/UserCouponRepository.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import shop.bluebooktle.backend.coupon.entity.UserCoupon;
-import shop.bluebooktle.backend.coupon.repository.query.CouponQueryRepository;
 import shop.bluebooktle.common.entity.auth.User;
 
 public interface UserCouponRepository extends JpaRepository<UserCoupon, Long>, CouponQueryRepository {
